pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coremake"
version = "0.1.0"
description = "CoRemake: Literature update tracking and logic chain extraction with Engram memory augmentation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "transformers>=4.36",
    "pydantic>=2.0",
    "PyYAML>=6.0",
    "pymupdf>=1.23",
    "networkx>=3.0",
    "faiss-cpu>=1.7",
    "scikit-learn>=1.3",
    "tqdm>=4.60",
    "jsonlines>=4.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "ipython"]
gpu = ["faiss-gpu>=1.7"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"
